some gcc warnings
git-svn-id: https://cgru.svn.sourceforge.net/svnroot/cgru@1375 0f86f178-d61e-0410-9d12-f6255c3815b8

+++ afanasy/trunk/src/libafanasy/msgclasses/mcgeneral.cpp
@@ -87,8 +87,8 @@ void MCGeneral::stdOut( bool full) const
 
    if( full == false ) return;
 
-   size_t count = list.size();
-   printf("ids[%u]=", count);
+   int count = int(list.size());
+   printf("ids[%d]=", count);
    for( unsigned p = 0; p < count; p++) printf(" %d", list[p]);
    printf(".\n");
 }
